Move supported_versions_for_module to tests._utils
